pull request clean-up
