+++ calico_versioned_docs/version-3.27/getting-started/index.mdx
@@ -2,20 +2,9 @@
 description: Install Calico on nodes and hosts for popular orchestrators, and install the calicoctl command line interface (CLI) tool.
 hide_table_of_contents: true
 ---
-
 # Install Calico
 
 import DocCardList from '@theme/DocCardList';
-import { DocCardLink, DocCardLinkLayout } from '../../../src/___new___/components';
 import { useCurrentSidebarCategory } from '@docusaurus/theme-common';
 
-### Test card layout
-
-Some text here explaining something about Calico
-
-<DocCardLinkLayout>
-  <DocCardLink docId='getting-started/bare-metal/requirements' />
-  <DocCardLink docId='getting-started/openstack/installation/verification' />
-  <DocCardLink docId='getting-started/openstack/installation/devstack' />
-  <DocCardLink docId='getting-started/kubernetes/hardway/istio-integration' />
-</DocCardLinkLayout>
+<DocCardList items={useCurrentSidebarCategory().items} />
